chore: renamed support_functions as utils
